ai-engine/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pdf2image==1.17.0
python-dotenv==1.0.0
Pillow==10.1.0
